requirements.txt
annotated-types==0.7.0
anthropic==0.50.0
anyio==4.9.0
argcomplete==3.6.2
asyncio==3.4.3
backoff==2.2.1
beautifulsoup4==4.13.4
boto3==1.38.6
botocore==1.38.6
browser-use==0.1.41
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
cohere==5.15.0
colorama==0.4.6
coloredlogs==15.0.1
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
distro==1.9.0
dnspython==2.7.0
dotenv==0.9.9
easyocr==1.7.2
email_validator==2.2.0
et_xmlfile==2.0.0
eval_type_backport==0.2.2
faiss-cpu==1.11.0
fastavro==1.10.0
fastembed==0.6.1
ffmpeg==1.4
filelock==3.18.0
filetype==1.2.0
flatbuffers==25.2.10
fsspec==2025.3.2
google-ai-generativelanguage==0.6.18
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
greenlet==3.2.1
griffe==1.7.3
groq==0.23.1
grpcio==1.72.0rc1
grpcio-status==1.72.0rc1
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
ImageHash==4.3.2
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.6.1
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.22
langchain-anthropic==0.3.3
langchain-aws==0.2.19
langchain-core==0.3.49
langchain-deepseek==0.1.3
langchain-google-genai==2.1.2
langchain-ollama==0.3.0
langchain-openai==0.3.11
langchain-text-splitters==0.3.7
langsmith==0.3.39
lazy_loader==0.4
llvmlite==0.44.0
logfire-api==3.14.1
loguru==0.7.3
lxml==5.4.0
markdown-it-py==3.0.0
markdownify==1.1.0
MarkupSafe==3.0.2
mcp==1.6.0
mdurl==0.1.2
mem0ai==0.1.88
mistralai==1.7.0
mmh3==5.1.0
monotonic==1.6
more-itertools==10.7.0
moviepy==2.1.2
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.4
numba==0.61.2
numpy==2.2.5
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
oauthlib==3.2.2
ollama==0.4.8
onnxruntime==1.21.1
openai==1.76.2
openai-whisper @ git+https://github.com/openai/whisper.git@517a43ecd132a2089d85f4ebc044728a71d49f6e
opencv-python-headless==4.11.0.86
opentelemetry-api==1.32.1
orjson==3.10.18
packaging==24.2
phonenumbers==9.0.4
pillow==10.4.0
playwright==1.52.0
portalocker==2.10.1
posthog==3.25.0
proglog==0.1.11
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==6.30.2
psutil==7.0.0
psycopg2-binary==2.9.10
py_rust_stemmers==0.1.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.3.0.post6
pydantic==2.10.6
pydantic-ai==0.1.8
pydantic-ai-slim==0.1.8
pydantic-evals==0.1.8
pydantic-graph==0.1.8
pydantic-settings==2.9.1
pydantic_core==2.27.2
pyee==13.0.0
Pygments==2.19.1
pyparsing==3.2.3
pyperclip==1.9.0
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-telegram-bot==22.0
pytz==2024.2
PyWavelets==1.8.0
PyYAML==6.0.2
qdrant-client==1.14.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rsa==4.9.1
s3transfer==0.12.0
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
screeninfo==0.8.1
sentence-transformers==4.1.0
setuptools==80.1.0
shapely==2.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.40
sse-starlette==2.3.3
starlette==0.46.2
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tifffile==2025.3.30
tika==3.1.0
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
transformers==4.51.3
triton==3.3.0
types-requests==2.32.0.20250328
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
wcwidth==0.2.13
wrapt==1.17.2
XlsxWriter==3.2.3
zipp==3.21.0
zstandard==0.23.0
